visualise graph
