+++ python/python/raphtory/__init__.py
@@ -1,0 +1,6 @@
+from .matplotlib import matplot_draw
+from .raphtory import *
+
+__doc__ = raphtory.__doc__
+if hasattr(raphtory, "__all__"):
+    __all__ = raphtory.__all__
