--- raphtory/Cargo.toml
@@ -1,42 +1,0 @@
-[package]
-name = "raphtory"
-description = "Python package for DocBrown, a temporal graph library"
-version = "0.0.6"
-edition.workspace = true
-rust-version.workspace = true
-keywords.workspace = true
-authors.workspace = true
-documentation.workspace = true
-repository.workspace = true
-license.workspace = true
-readme.workspace = true
-homepage.workspace = true
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[lib]
-name = "raphtory"
-crate-type = ["cdylib"]
-
-[dependencies]
-pyo3 =  "0.18.1"
-docbrown-db = {path = "../docbrown/db", version = "0.0.6" }
-docbrown-core = {path = "../docbrown/core", version = "0.0.6" }
-
-parking_lot = { version = "0.12" , features = ["serde"] }
-flume = "0.10"
-futures = {version = "0.3", features = ["thread-pool"] }
-replace_with = "0.1"
-itertools="0.10"
-csv = "1.1.6"
-flate2 = "1.0"
-regex = "1"
-serde = { version = "1", features = ["derive", "rc"] }
-rayon = "1"
-chrono = "0.4"
-bincode = "1"
-
-[features]
-extension-module = ["pyo3/extension-module"]
-default = ["extension-module"]
-
-[build-dependencies]
-pyo3-build-config = "0.18.1"
