--- raphtory/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    pyo3_build_config::add_extension_module_link_args();
-}
