--- raphtory/pyproject.toml
@@ -1,16 +1,0 @@
-[build-system]
-requires = ["maturin>=0.13,<0.14"]
-build-backend = "maturin"
-
-[project]
-name = "raphtory"
-requires-python = ">=3.7"
-classifiers = [
-    "Programming Language :: Rust",
-    "Programming Language :: Python :: Implementation :: CPython",
-    "Programming Language :: Python :: Implementation :: PyPy",
-]
-
-[tool.maturin]
-features = ["pyo3/extension-module"]
-
